requirements.txt
newsapi-python==0.2.7
python-dotenv==1.1.1
newspaper3k==0.2.8
google-generativeai==0.8.3
nltk>=3.8.1
